ui is now resizable and fullscreenable

+++ Hedra/Engine/Player/Inventory/InventoryArrayInterface.cs
@@ -43,7 +43,7 @@ namespace Hedra.Engine.Player.Inventory
             var offset = new Vector2(size.X, size.Y);
             var wholeSize = new Vector2(
                 size.X * (_length - 1 - (_length - 1) / SlotsPerLine * SlotsPerLine),
-                size.Y * ((_length - 1) / SlotsPerLine)
+                size.Y * ((_length - 1) / (float)SlotsPerLine)
             );
             for (var i = 0; i < _length; i++)
             {
